added fixes for the oidc header and some refactoring of that area.
